requirements.txt
streamlit
snowflake-sqlalchemy==1.4.7
llama_index==0.7.9
